+++ src/test/rv/new-request-no-asid.c
@@ -1,4 +1,4 @@
-#include "src/main/c/xfiles.h"
+#include "src/main/c/xfiles-user.h"
 
 int main() {
   new_write_request(0, 0, 0);
