+++ src/test/rv/write-register-no-asid.c
@@ -1,4 +1,4 @@
-#include "src/main/c/xfiles.h"
+#include "src/main/c/xfiles-user.h"
 
 int main() {
   write_register(0, 0, 0);
