+++ tools/common/Makefrag-rv
@@ -24,27 +24,38 @@ vpath %.c $(DIR_TOP)/src/test/rv
 $(DIR_BUILD)/newlib/%.o: %.c | $(DIR_BUILD)/newlib
 	$(RV_CC) -Wall -Werror -march=RV64IMAFDXcustom -I. -c $< -o $@
 
-$(DIR_BUILD)/newlib/libxfiles.a: $(XFILES_LIBRARIES_OBJECTS_NEWLIB)
-	$(RV_AR) rcs $@ $(XFILES_LIBRARIES_OBJECTS_NEWLIB)
+$(DIR_BUILD)/newlib/libxfiles-%.a: $(DIR_BUILD)/newlib/xfiles-%.o
+	$(RV_AR) rcs $@ $<
 
 $(DIR_BUILD)/linux/%.o: %.c | $(DIR_BUILD)/linux
 	$(RV_CC_LINUX) -Wall -Werror -march=RV64IMAFDXcustom -I. -c $< -o $@
 
-$(DIR_BUILD)/linux/libxfiles.a: $(XFILES_LIBRARIES_OBJECTS_LINUX)
-	$(RV_AR_LINUX) rcs $@ $(XFILES_LIBRARIES_OBJECTS_LINUX)
+$(DIR_BUILD)/linux/libxfiles-%.a: $(DIR_BUILD)/linux/xfiles-%.o
+	$(RV_AR_LINUX) rcs $@ $<
 
 #------------------- RISC-V Tests
+
+RV_CFLAGS=-Wall -Werror -static -march=RV64IMAFDXcustom -I. -I$(DIR_BUILD)/nets
+RV_NEWLIB_LFLAGS=-L$(DIR_TOP)/usr/lib-rv-newlib -lxfiles-user
+RV_LINUX_LFLAGS=-L$(DIR_TOP)/usr/lib-rv-linux -lxfiles-user
+
 $(DIR_BUILD)/newlib/fann-soft.rv: fann-soft.c $(XFILES_LIBRARIES_NEWLIB) $(DIR_BUILD)/fann-rv-newlib/libfann.a
-	$(RV_CC) -Wall -Werror -static -march=RV64IMAFDXcustom -I. -I$(DIR_BUILD)/nets -I$(DIR_TOP)/usr/include $< -o $@ -L$(DIR_TOP)/usr/lib-rv-newlib -lxfiles -lfann -lm
+	$(RV_CC) $(RV_CFLAGS) $< -o $@ $(RV_NEWLIB_LFLAGS) -lfann -lm
 
 $(DIR_BUILD)/linux/fann-soft.rv: fann-soft.c $(DIR_BUILD_NETS) $(XFILES_LIBRARIES_LINUX) $(DIR_BUILD)/linux $(DIR_BUILD)/fann-rv-linux/libfann.a
-	$(RV_CC_LINUX) -Wall -Werror -static -march=RV64IMAFDXcustom -I. -I$(DIR_BUILD_NETS) -I$(DIR_TOP)/usr/include $< -o $@ -L$(DIR_TOP)/usr/lib-rv-linux -lxfiles -lfann -lm
+	$(RV_CC_LINUX) $(RV_CFLAGS) $< -o $@ $(RV_LINUX_LFLAGS) -lfann -lm
+
+$(DIR_BUILD)/newlib/fann-xfiles.rv: fann-xfiles.c $(XFILES_LIBRARIES_NEWLIB) $(DIR_BUILD)/fann-rv-newlib/libfann.a
+	$(RV_CC) $(RV_CFLAGS) $< -o $@ $(RV_NEWLIB_LFLAGS) -lfixedfann -lm
+
+$(DIR_BUILD)/linux/fann-xfiles.rv: fann-xfiles.c $(XFILES_LIBRARIES_LINUX) $(DIR_BUILD)/fann-rv-linux/libfann.a
+	$(RV_CC_LINUX) $(RV_CFLAGS) $< -o $@ $(RV_LINUX_LFLAGS) -lfixedfann -lm
 
 $(DIR_BUILD)/newlib/%.rv: %.c $(XFILES_LIBRARIES_NEWLIB) $(DIR_BUILD)/fann-rv-newlib/libfann.a | $(DIR_BUILD)/newlib
-	$(RV_CC) -Wall -Werror -static -march=RV64IMAFDXcustom -I. -I$(DIR_BUILD_NETS) -I$(DIR_TOP)/usr/include $< -o $@ -L$(DIR_TOP)/usr/lib-rv-newlib -lxfiles -lfixedfann -lm
+	$(RV_CC) $(RV_CFLAGS) $< -o $@ $(RV_NEWLIB_LFLAGS)
 
 $(DIR_BUILD)/linux/%.rv: %.c $(XFILES_LIBRARIES_LINUX) $(DIR_BUILD)/fann-rv-linux/libfann.a | $(DIR_BUILD)/linux
-	$(RV_CC_LINUX) -Wall -Werror -static -march=RV64IMAFDXcustom -I. -I$(DIR_BUILD_NETS) -I$(DIR_TOP)/usr/include $< -o $@ -L$(DIR_TOP)/usr/lib-rv-linux -lxfiles -lfixedfann -lm
+	$(RV_CC_LINUX) $(RV_CFLAGS) $< -o $@ $(RV_LINUX_LFLAGS)
 
 $(DIR_BUILD)/newlib/%.rvS: $(DIR_BUILD)/newlib/%.rv
 	$(RV_OBJDUMP) -S $< > $@
