--- usr/lib-rv-linux/libxfiles.a
@@ -1,1 +1,0 @@
-../../build/linux/libxfiles.a
