+++ usr/lib-rv-newlib/libxfiles-supervisor.a
@@ -1,0 +1,1 @@
+../../build/newlib/libxfiles-supervisor.a
