--- usr/lib-rv-newlib/libxfiles.a
@@ -1,1 +1,0 @@
-../../build/newlib/libxfiles.a
